Attempt storing complex maps of data as value in BaseMeasurement child.
This isn't working very well. Why bother having the Map as the unit? Why not let
the measurement class just define it's own arbitrary inner types and methods if
it needs to? The problem is the requirement for a single template type from
BaseMeasurement. OK, what if we just implement Measurement? Well, we depend on
BaseMeasurment in a bunch of places (e.g. for construction of a Measurement
object from RawMeasurement), so that has to be refactored before we can make
Measurement children that don't use BaseMeasurement.

An alternative is to not have basemeasurement worry about the type and value,
and leave that to another sub-class. The basemeasurement could have the
constructors and timestamp stuff and that's it.